[#902] 외부 캘린더 상세의 편집 입력을 공용 컴포넌트로 추출
- EventNameInputView 신설 — 색 바(render-prop) + 큰 제목 TextField, 부모 @FocusState 바인딩 수용
- EventTimeSelectView 신설 — 시각 선택 + 종일 토글 + 인라인 DatePicker, TimeSelecting enum·피커 토글 @State 내부 소유, onBeginSelecting 훅으로 부모 텍스트 포커스 해제
- EventTextInputRow 신설 — 아이콘 + 한 줄 TextField(장소 등 공용)
- EventMemoInputView 신설 — doc.text 아이콘 + placeholder 겹친 TextEditor
- GoogleCalendarEventDetailView.nameInputView/timeSelectView/locationInputView/memoInputView — 위 컴포넌트 호출로 교체, 이관된 private 멤버(TimeSelecting enum·isTimeSelecting·SelectedTime.startDate/endDate 등) 제거
- EventDetailScene/CLAUDE.md 프레임워크 스코프 컴포넌트 표에 4개 등재
