+++ Presentations/EventDetailScene/CLAUDE.md
@@ -169,6 +169,10 @@ graph TD
 | `EventTimeTextView` | `SelectTimeText`(연/일/시각) 라벨. `textColor`(selecting 하이라이트)·`isStrikethrough`(invalid)·`dayLineLimit` 파라미터로 화면별 변형 흡수 | EventDetailView, GoogleCalendarEventDetailView, AppleCalendarEventDetailView, DoneTodoDetailView |
 | `LandmarkLabelView` | 장소 라벨 (이름+주소+xmark 아이콘). EventDetailView는 Menu(삭제 액션)로 감싸 사용 | EventDetailView, DoneTodoDetailView |
 | `MoreActionMenuLabel` | 하단 더보기(ellipsis) 버튼 라벨 — 20×20 아이콘 + `Radius.regular` `secondaryBtnBackground` 배경. `Menu`의 label로만 사용하고 액션 목록은 호출부가 구성 | EventDetailView, HolidayEventDetailView, GoogleCalendarEventDetailView |
+| `EventNameInputView` | 선행 색 바(`colorBar` render-prop) + 큰 제목 TextField. 포커스는 부모 `@FocusState` 바인딩(`@FocusState.Binding`)을 받아 스크롤 연동 유지 | GoogleCalendarEventDetailView |
+| `EventTimeSelectView` | clock 아이콘 + 선택 시각(`EventTimeTextView` 재사용, 탭하면 인라인 DatePicker 토글) + 종일 토글. 피커 토글 상태는 내부 `@State` 소유, `onBeginSelecting` 훅으로 부모가 텍스트 포커스 해제 | GoogleCalendarEventDetailView |
+| `EventTextInputRow` | 아이콘 + 한 줄 TextField (장소·URL 등 공용). 포커스는 부모 `@FocusState` 바인딩 | GoogleCalendarEventDetailView |
+| `EventMemoInputView` | doc.text 아이콘 + placeholder 겹친 TextEditor. 포커스는 부모 `@FocusState` 바인딩 | GoogleCalendarEventDetailView |
 
 `GuideView/`의 가이드 오버레이 2종(ForemostEventGuideView·TodoEventGuideView)은 컴포넌트 패밀리로 별도 그룹.
